should be posixly correct now re option arguments
POSIX says, the option argument may follow either in the following
argv[] element or directly behind the option letter in the same
argv[] element. This also means that if options are clustered,
everything behind an option that takes the argument must be the
argument and not more option letters in that cluster.

The before behaviour allowed all option letters being clustered,
and needed option arguments were taken from the following argv[]
elements, but now the behaviour is (or should be, rather) posixly
correct in this respect. Long option behaviour isn't affected.

This was surprisingly easy to knit in, but it took me a while to
get it back under 10 kLoC again. This is, though, not something I
could take away.

Re the "should be" above: writing the tests is, again, very very
tedious. Why wasn't that (so much) the case with lingo?