make callable as default value user-accessible
They type of an option may be None and the default value a callable,
in which case no argument will be pulled from argv[], but instead
the default value will be called, and the result value will be used,
if it is non None. (In that case, an argument will still be pulled.)

While this is just an artifact of the mechanism the -h and -? default
options are implemented, it may well be useful in its own right.
Still, I'll rather keep it undocumented for the moment.