have the README ToC created again
For some weird and totally annoying reason, gh-md-toc doesn't create
a toc from the README source if there is a "-*- text -*-" on top
of the file. WTF??

As it does if the tag is "-*- major-mode: text -*-", this is not
only weird, but also stupid. Meaning it isn't Markdown if I want
to edit it in text-mode? Why not stalk down all other options to
set text mode, too? Argh.