testing secrets seems fine
In that course, revamp of put/getsecret with a common secrets file
reader (finally!). Code is much better now. Not as dead simple any
more than the original getsecret() implementation, but the backend,
meaning the file format, nearly is, except for a colon more (in
absence of options). *If* there are options, it is slightly more
complex, but still comfortably editable by hand except for the
encoding.