have include.py process shell commands, too
This is done by a "#include !<command>" directive, meaning if the
include argument begins with a bang, it is run as a shell command
and the output is included.

This enables significant simplification of the Makefile for building
the pgetopts README. Also, it feels like the right thing to do.