Table: improve alignment
The `align` parameter may, as before, be a string or None (or
unspecified), but also a sequence of those. In that case, the first
element is used for the first line of the tables, the other (if
present, or the first) for following lines.

If the alignment string has at least two characters and the last
one is an asterisk, the preceding character is used for all following
columns. Example: "llcr*" means the first two columns will be aligned
left, the third centered, and the fourth and all following will be
aligned right.